[FIX] UI 연결 작업 진행중
